+++ src/GeminiMigrationManager.h
@@ -31,7 +31,7 @@ class GeminiMigrationManager : Dispatch::Poller {
 
     bool lookupPriorityHashes(uint64_t hash);
 
-    uint64_t updateRegularPullProgress(uint64_t i);
+    uint64_t updateRegularPullProgress(uint32_t i);
 
   PRIVATE:
     // Shared RAMCloud information.
